backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
azure-cognitiveservices-speech>=1.38.0
pytest>=8.2.0
torch>=2.5.1
torchaudio>=2.5.1
speechbrain>=1.0.3
numpy>=1.26.0
soundfile>=0.12.1
python-multipart>=0.0.9
pydub>=0.25.1
python-docx>=1.1.0
python-pptx>=0.6.21
PyPDF2>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
